Split /utils into( /gtp, /options), and /test/utils ( /test/base_server_packets, /test/fake_modules, /test/module_loading, /test/utils) (#342)
* move gtp helpers from /utils to /gtp_helpers

* rename /utils to /options_helpers

* move base server packets to /test/base_server_packets

* move module loading helpers to /test/module_loading_helpers

* move fake modules to /test/fake_modules

* rename _helpers to ""

* add /gtp/** and /options/** to npm package.json for an npm install

* fix console logging version 6.0 to 6.0.0

* version bump from 6.0.0 to 6.0.1
