+++ game.js
@@ -1,7 +1,7 @@
 // vim: tw=120 softtabstop=4 shiftwidth=4
 
-const { decodeMoves } = require("./utils/decodeMoves");
-const { move2gtpvertex } = require("./utils/move2gtpvertex");
+const { decodeMoves } = require("./gtp/decodeMoves");
+const { move2gtpvertex } = require("./gtp/move2gtpvertex");
 
 const Bot = require('./bot').Bot;
 const console = require('./console').console;
