+++ gtp/char2num.js
