+++ gtp/decodeMoves.js
