+++ gtp/gtpchar2num.js
