+++ gtp/move2gtpvertex.js
