+++ gtp/num2char.js
