+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "gtp2ogs",
-  "version": "6.0.0",
+  "version": "6.0.1",
   "description": "Wrapper to allow Gnu Go Text Protocol speaking Go engines to connect to Online-Go.com and play games",
   "main": "gtp2ogs.js",
   "files": [
@@ -9,7 +9,10 @@
     "connection.js",
     "console.js",
     "game.js",
-    "gtp2ogs.js"
+    "gtp2ogs.js",
+    "pv.js",
+    "gtp/**",
+    "options/**"
   ],
   "bin": {
     "gtp2ogs": "gtp2ogs.js"
