+++ test/checkChallenge.test.js
@@ -4,11 +4,11 @@ const assert = require('assert');
 const https = require('https');
 const sinon = require('sinon');
 
-const { base_challenge } = require('./utils/base_challenge');
-const { FakeAPI } = require('./utils/FakeAPI');
-const { FakeSocket } = require('./utils/FakeSocket');
-const { getNewConfigUncached } = require('./utils/getNewConfigUncached');
-const { getNewConnectionUncached } = require('./utils/getNewConnectionUncached');
+const { base_challenge } = require('./base_server_packets/base_challenge');
+const { FakeAPI } = require('./fake_modules/FakeAPI');
+const { FakeSocket } = require('./fake_modules/FakeSocket');
+const { getNewConfigUncached } = require('./module_loading/getNewConfigUncached');
+const { getNewConnectionUncached } = require('./module_loading/getNewConnectionUncached');
 const { stub_console } = require('./utils/stub_console');
 
 let config;
