+++ test/fake_modules/FakeAPI.js
