+++ test/fake_modules/FakeGTP.js
