+++ test/fake_modules/FakeSocket.js
