+++ test/module_loading/assignConfigArguments.js
