+++ test/module_loading/getNewConfig.js
