+++ test/module_loading/getNewConfigUncached.js
