+++ test/module_loading/getNewConnectionUncached.js
