+++ test/module_loading/requireUncached.js
